+++ src/lib/link-meta/link-meta.ts
@@ -1,10 +1,9 @@
 import {BskyAgent} from '@atproto/api'
 
-import {LINK_META_PROXY} from 'lib/constants'
-import {getGiphyMetaUri} from 'lib/strings/embed-player'
-import {parseStarterPackUri} from 'lib/strings/starter-pack'
+import {LINK_META_PROXY} from '#/lib/constants'
+import {getGiphyMetaUri} from '#/lib/strings/embed-player'
+import {parseStarterPackUri} from '#/lib/strings/starter-pack'
 import {isBskyAppUrl} from '../strings/url-helpers'
-import {extractBskyMeta} from './bsky'
 
 export enum LikelyType {
   HTML,
@@ -31,7 +30,10 @@ export async function getLinkMeta(
   timeout = 15e3,
 ): Promise<LinkMeta> {
   if (isBskyAppUrl(url) && !parseStarterPackUri(url)) {
-    return extractBskyMeta(agent, url)
+    return {
+      likelyType: LikelyType.AtpData,
+      url,
+    }
   }
 
   let urlp
